Took a pass at the coral intake logic to change it up. Logic:
if we want to eject, we do. If we have a gamepiece, we hold.
Changed: Next, if we have a coral in front and back sensor (so it's close to being in right position), we move it forward.
Then, if we are just intaking (without seeing front sensor), it intakes much faster.
Otherwise, the voltage is 0.

It makes sense to me, but it's untested.